requirements.txt
PyQt5==5.15.9
opencv-python==4.8.1.78
numpy==1.24.3
mysql-connector-python==8.1.0
insightface==0.7.3
pickle5==0.0.11
python-dateutil==2.8.2